+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void TakeDamage(int damageAmount);
+    void TakeDamage(float damageAmount);
 }
